move files to appropriate packages
