+++ api/src/main/kotlin/com/intellij/vim/api/ThinApi.kt
@@ -8,30 +8,10 @@
 
 package com.intellij.vim.api
 
-val TextType.isLine: Boolean
-  get() = this == TextType.LINE_WISE
-
 enum class TextType {
   CHARACTER_WISE,
   LINE_WISE,
   BLOCK_WISE,
-}
-
-sealed interface Range {
-  data class Simple(val start: Int, val end: Int): Range
-
-  data class Block(val ranges: Array<Simple>): Range {
-    override fun equals(other: Any?): Boolean {
-      if (this === other) return true
-      if (javaClass != other?.javaClass) return false
-      other as Block
-      return ranges.contentEquals(other.ranges)
-    }
-
-    override fun hashCode(): Int {
-      return ranges.contentHashCode()
-    }
-  }
 }
 
 data class Line(val number: Int, val text: String, val start: Int, val end: Int)
@@ -47,23 +27,3 @@ data class CaretInfo(
 )
 
 interface HighlightId
-
-data class Color(
-  val hexCode: String
-) {
-  constructor(r: Int, g: Int, b: Int, a: Int = 255): this(String.format("#%02x%02x%02x%02x", r, g, b, a))
-
-  val r: Int = hexCode.substring(1..2).toInt(16)
-
-  val g: Int = hexCode.substring(3..4).toInt(16)
-
-  val b: Int = hexCode.substring(5..6).toInt(16)
-
-  val a: Int = if (hexCode.length == 9) hexCode.substring(7..8).toInt(16) else 255
-
-  init {
-    require(hexCode.matches(Regex("^#[0-9A-Fa-f]{6}([0-9A-Fa-f]{2})?$"))) {
-      "Hex code should be in format #RRGGBB[AA]"
-    }
-  }
-}
